completed

+++ main.js
@@ -5,12 +5,12 @@ function submitIssue(e) {
   const description = getInputValue('issueDescription');
   const severity = getInputValue('issueSeverity');
   const assignedTo = getInputValue('issueAssignedTo');
-  const id = Math.floor(Math.random()*100000000) + '';
+  const id = Math.floor(Math.random() * 100000000) + '';
   const status = 'Open';
 
   const issue = { id, description, severity, assignedTo, status };
   let issues = [];
-  if (localStorage.getItem('issues')){
+  if (localStorage.getItem('issues')) {
     issues = JSON.parse(localStorage.getItem('issues'));
   }
   issues.push(issue);
@@ -30,9 +30,13 @@ const closeIssue = id => {
 }
 
 const deleteIssue = id => {
-  const issues = JSON.parse(localStorage.getItem('issues'));
-  const remainingIssues = issues.filter( issue.id !== id )
-  localStorage.setItem('issues', JSON.stringify(remainingIssues));
+  const isConfirm = confirm('Are you sure to delete this issues')
+  if (isConfirm) {
+    const issues = JSON.parse(localStorage.getItem('issues'));
+    const remainingIssues = issues.filter(issue => issue.id != id)
+    localStorage.setItem('issues', JSON.stringify(remainingIssues));
+    fetchIssues();
+  }
 }
 
 const fetchIssues = () => {
@@ -41,9 +45,9 @@ const fetchIssues = () => {
   issuesList.innerHTML = '';
 
   for (var i = 0; i < issues.length; i++) {
-    const {id, description, severity, assignedTo, status} = issues[i];
+    const { id, description, severity, assignedTo, status } = issues[i];
 
-    issuesList.innerHTML +=   `<div class="well">
+    issuesList.innerHTML += `<div class="well">
                               <h6>Issue ID: ${id} </h6>
                               <p><span class="label label-info"> ${status} </span></p>
                               <h3> ${description} </h3>
